pyproject.toml
[project]
name = "marketing-segmentation-poc"
version = "0.1.0"
description = ""
authors = [
    {name = "Ruslan Mikhailov",email = "bkr0308@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "openai (>=2.7.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "redshift-connector (>=2.1.9,<3.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "psycopg[binary] (>=3.2.13,<4.0.0)",
    "pytest-env (>=1.2.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
start = "uvicorn.main:run"